GooglePlacesChacheRepoImpl ² GoogleListRestaurantUseCase are Singleton. Add PlacesSearchApiResponseAdapter with ViewHolder, Not finished yet. add updateUI in GoogleMapFragment with nearby Restaurants.

+++ app/src/main/java/com/gt/go4lunch/data/Location.kt
@@ -7,9 +7,9 @@ class Location {
 
     @SerializedName("lat")
     @Expose
-    var lat: Float? = null
+    var lat: Double? = null
     @SerializedName("lng")
     @Expose
-    var lng: Float? = null
+    var lng: Double? = null
 
 }
